fix: upcast to float32 for LoRA matmul, cast back to input dtype
T4 GPU does not support native bfloat16 matmul.
Casting A/B to bfloat16 still fails -- A and B must stay float32.
Upcast x to float32, compute, then downcast result to x.dtype.

+++ src/core/lora.py
@@ -14,7 +14,8 @@ class LoRALayer(nn.Module):
         self.B = nn.Parameter(torch.zeros(rank, out_features))
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
-        return (x @ self.A.to(x.dtype) @ self.B.to(x.dtype)) * self.scaling
+        # T4 GPU bfloat16 matmul desteklemiyor -- float32'de hesapla, geri cast et
+        return ((x.float() @ self.A @ self.B) * self.scaling).to(x.dtype)
 
 
 class LinearWithLoRA(nn.Module):
